+++ tuxbox/neutrino/src/gui/pictureviewer.cpp
@@ -167,7 +167,7 @@ int CPictureViewerGui::show()
 	CLCD::getInstance()->setMode(CLCD::MODE_MENU_UTF8, g_Locale->getText("pictureviewer.head") );
 	m_state=MENU;
 
-	uint msg; uint64_t data;
+	uint msg; uint data;
 	int timeout;
 
 	bool loop=true;
